+++ htdocs/concatpdf/core/modules/modConcatPdf.class.php
@@ -52,11 +52,11 @@ class modConcatPdf extends DolibarrModules
 		// Module label (no space allowed), used if translation string 'ModuleXXXName' not found (where XXX is value of numeric property 'numero' of module)
 		$this->name = preg_replace('/^mod/i','',get_class($this));
 		// Module description used if translation string 'ModuleXXXDesc' not found (XXX is id value)
-		$this->description = "Concat pdfs found into a directory to generated pdf files (proposals, orders, invoices)";
+		$this->description = "Concat on or several PDFs found into a directory to the generated PDF files (proposals, orders, invoices)";
         $this->editor_name = 'NLTechno';
         $this->editor_url = 'https://www.nltechno.com';
 		// Possible values for version are: 'development', 'experimental', 'dolibarr' or version
-		$this->version = '6.0.1';
+		$this->version = '6.0.2';
 		// Key used in llx_const table to save module status enabled/disabled (XXX is id value)
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		// Name of png file (without png) used for this module